apps/api/requirements.txt
fastapi>=0.115.0
uvicorn>=0.30.0
pydantic>=2.7.0
supabase>=2.4.0
python-dotenv>=1.0.0
pytest>=8.0.0
httpx>=0.27.0
google-generativeai
